Pre-submission tweaks.
"Thus we ensure that parenthesized, hierarchical structures appear in
programs next to instructions that can make use of them and not next to
instructions that cannot." -> "Thus we make it more likely that
parenthesized, hierarchical structures appear in programs next to
instructions that can make use of them." We do not ensure that they
cannot appear next to instructions that won't make use of them, and in
fact we provide a way to make that happen (noop_open_paren) because
that might be a good idea in the context of exotic code manipulation.
And in fact we don't ensure it the other way either, because of
noop_delete_prev_paren_pair.

Tweaked paragraph on "epigenetic" terminology to make more clear what's
describing the prior work and what's describing the current work.

Unescaped underscores in verbatim code.